cheri: Fix compile and warnings in net/sched/
Add necessary casts and change unsigned longs to uintptr_t
as required.

+++ net/sched/cls_u32.c
@@ -720,7 +720,7 @@ static void u32_unbind_filter(struct tcf_proto *tp, struct tc_u_knode *n,
 }
 
 static void u32_bind_filter(struct tcf_proto *tp, struct tc_u_knode *n,
-			    unsigned long base, struct nlattr **tb)
+			    uintptr_t base, struct nlattr **tb)
 {
 	if (tb[TCA_U32_CLASSID]) {
 		n->res.classid = nla_get_u32(tb[TCA_U32_CLASSID]);
@@ -858,7 +858,7 @@ static struct tc_u_knode *u32_init_knode(struct net *net, struct tcf_proto *tp,
 }
 
 static int u32_change(struct net *net, struct sk_buff *in_skb,
-		      struct tcf_proto *tp, unsigned long base, u32 handle,
+		      struct tcf_proto *tp, uintptr_t base, u32 handle,
 		      struct nlattr **tca, void **arg, u32 flags,
 		      struct netlink_ext_ack *extack)
 {
@@ -1318,8 +1318,8 @@ static int u32_reoffload(struct tcf_proto *tp, bool add, flow_setup_cb_t *cb,
 	return 0;
 }
 
-static void u32_bind_class(void *fh, u32 classid, unsigned long cl, void *q,
-			   unsigned long base)
+static void u32_bind_class(void *fh, u32 classid, uintptr_t cl, void *q,
+			   uintptr_t base)
 {
 	struct tc_u_knode *n = fh;
