+++ net/sched/sch_drr.c
@@ -51,7 +51,7 @@ static const struct nla_policy drr_policy[TCA_DRR_MAX + 1] = {
 };
 
 static int drr_change_class(struct Qdisc *sch, u32 classid, u32 parentid,
-			    struct nlattr **tca, unsigned long *arg,
+			    struct nlattr **tca, uintptr_t *arg,
 			    struct netlink_ext_ack *extack)
 {
 	struct drr_sched *q = qdisc_priv(sch);
@@ -132,7 +132,7 @@ static int drr_change_class(struct Qdisc *sch, u32 classid, u32 parentid,
 
 	qdisc_class_hash_grow(sch, &q->clhash);
 
-	*arg = (unsigned long)cl;
+	*arg = (uintptr_t)cl;
 	return 0;
 }
 
@@ -143,7 +143,7 @@ static void drr_destroy_class(struct Qdisc *sch, struct drr_class *cl)
 	kfree(cl);
 }
 
-static int drr_delete_class(struct Qdisc *sch, unsigned long arg,
+static int drr_delete_class(struct Qdisc *sch, uintptr_t arg,
 			    struct netlink_ext_ack *extack)
 {
 	struct drr_sched *q = qdisc_priv(sch);
@@ -165,12 +165,12 @@ static int drr_delete_class(struct Qdisc *sch, unsigned long arg,
 	return 0;
 }
 
-static unsigned long drr_search_class(struct Qdisc *sch, u32 classid)
+static uintptr_t drr_search_class(struct Qdisc *sch, u32 classid)
 {
-	return (unsigned long)drr_find_class(sch, classid);
+	return (uintptr_t)drr_find_class(sch, classid);
 }
 
-static struct tcf_block *drr_tcf_block(struct Qdisc *sch, unsigned long cl,
+static struct tcf_block *drr_tcf_block(struct Qdisc *sch, uintptr_t cl,
 				       struct netlink_ext_ack *extack)
 {
 	struct drr_sched *q = qdisc_priv(sch);
@@ -183,7 +183,7 @@ static struct tcf_block *drr_tcf_block(struct Qdisc *sch, unsigned long cl,
 	return q->block;
 }
 
-static unsigned long drr_bind_tcf(struct Qdisc *sch, unsigned long parent,
+static uintptr_t drr_bind_tcf(struct Qdisc *sch, uintptr_t parent,
 				  u32 classid)
 {
 	struct drr_class *cl = drr_find_class(sch, classid);
@@ -191,17 +191,17 @@ static unsigned long drr_bind_tcf(struct Qdisc *sch, unsigned long parent,
 	if (cl)
 		qdisc_class_get(&cl->common);
 
-	return (unsigned long)cl;
+	return (uintptr_t)cl;
 }
 
-static void drr_unbind_tcf(struct Qdisc *sch, unsigned long arg)
+static void drr_unbind_tcf(struct Qdisc *sch, uintptr_t arg)
 {
 	struct drr_class *cl = (struct drr_class *)arg;
 
 	qdisc_class_put(&cl->common);
 }
 
-static int drr_graft_class(struct Qdisc *sch, unsigned long arg,
+static int drr_graft_class(struct Qdisc *sch, uintptr_t arg,
 			   struct Qdisc *new, struct Qdisc **old,
 			   struct netlink_ext_ack *extack)
 {
@@ -218,21 +218,21 @@ static int drr_graft_class(struct Qdisc *sch, unsigned long arg,
 	return 0;
 }
 
-static struct Qdisc *drr_class_leaf(struct Qdisc *sch, unsigned long arg)
+static struct Qdisc *drr_class_leaf(struct Qdisc *sch, uintptr_t arg)
 {
 	struct drr_class *cl = (struct drr_class *)arg;
 
 	return cl->qdisc;
 }
 
-static void drr_qlen_notify(struct Qdisc *csh, unsigned long arg)
+static void drr_qlen_notify(struct Qdisc *csh, uintptr_t arg)
 {
 	struct drr_class *cl = (struct drr_class *)arg;
 
 	list_del(&cl->alist);
 }
 
-static int drr_dump_class(struct Qdisc *sch, unsigned long arg,
+static int drr_dump_class(struct Qdisc *sch, uintptr_t arg,
 			  struct sk_buff *skb, struct tcmsg *tcm)
 {
 	struct drr_class *cl = (struct drr_class *)arg;
@@ -254,7 +254,7 @@ nla_put_failure:
 	return -EMSGSIZE;
 }
 
-static int drr_dump_class_stats(struct Qdisc *sch, unsigned long arg,
+static int drr_dump_class_stats(struct Qdisc *sch, uintptr_t arg,
 				struct gnet_dump *d)
 {
 	struct drr_class *cl = (struct drr_class *)arg;
@@ -285,7 +285,7 @@ static void drr_walk(struct Qdisc *sch, struct qdisc_walker *arg)
 
 	for (i = 0; i < q->clhash.hashsize; i++) {
 		hlist_for_each_entry(cl, &q->clhash.hash[i], common.hnode) {
-			if (!tc_qdisc_stats_dump(sch, (unsigned long)cl, arg))
+			if (!tc_qdisc_stats_dump(sch, (uintptr_t)cl, arg))
 				return;
 		}
 	}
